+++ service-model-generator/modules/service-model-generator-ls-extension/src/main/java/io/ballerina/servicemodelgenerator/extension/util/Utils.java
@@ -50,8 +50,6 @@ import io.ballerina.compiler.syntax.tree.SyntaxKind;
 import io.ballerina.compiler.syntax.tree.Token;
 import io.ballerina.compiler.syntax.tree.TypeDefinitionNode;
 import io.ballerina.compiler.syntax.tree.TypeDescriptorNode;
-import io.ballerina.modelgenerator.commons.Annotation;
-import io.ballerina.modelgenerator.commons.ServiceDatabaseManager;
 import io.ballerina.projects.Document;
 import io.ballerina.servicemodelgenerator.extension.ServiceModelGeneratorConstants;
 import io.ballerina.servicemodelgenerator.extension.model.Codedata;
@@ -170,21 +168,6 @@ public final class Utils {
         }
     }
 
-    public static void enableContractFirstApproach(Service service) {
-        Value designApproach = service.getDesignApproach();
-        if (Objects.nonNull(designApproach) && Objects.nonNull(designApproach.getChoices())
-                && !designApproach.getChoices().isEmpty()) {
-            designApproach.getChoices().forEach(choice -> choice.setEnabled(false));
-            designApproach.getChoices().stream()
-                    .filter(choice -> choice.getMetadata().label().equals("Import From OpenAPI Specification"))
-                    .findFirst()
-                    .ifPresent(approach -> {
-                approach.setEnabled(true);
-                approach.getProperties().remove("spec");
-            });
-        }
-    }
-
     private static Optional<Service> getServiceByServiceType(String serviceType) {
         InputStream resourceStream = Utils.class.getClassLoader()
                 .getResourceAsStream(String.format("services/%s.json", serviceType.replaceAll(":", ".")));
@@ -206,47 +189,6 @@ public final class Utils {
         }
         ExpressionNode expressionNode = expressions.get(0);
         return Optional.of(expressionNode);
-    }
-
-    public static Service fromHttpServiceWithContract(TypeDefinitionNode serviceTypeNode,
-                                                      ServiceDeclarationNode serviceDeclarationNode,
-                                                      SemanticModel semanticModel) {
-        Service serviceModel = Service.getNewService();
-        Value serviceContractType = new Value.ValueBuilder()
-                .enabled(true)
-                .valueType(ServiceModelGeneratorConstants.VALUE_TYPE_IDENTIFIER)
-                .value(serviceTypeNode.typeName().text().trim())
-                .build();
-        serviceModel.setServiceContractTypeName(serviceContractType);
-        List<Function> functionModels = new ArrayList<>();
-        serviceDeclarationNode.members().forEach(member -> {
-            if (member instanceof FunctionDefinitionNode functionDefinitionNode) {
-                Function functionModel = getFunctionModel(functionDefinitionNode, semanticModel, true,
-                        false, Map.of());
-                functionModels.add(functionModel);
-            }
-        });
-        serviceModel.setFunctions(functionModels);
-        return serviceModel;
-    }
-
-    public static Service getServiceModel(ServiceDeclarationNode serviceDeclarationNode, SemanticModel semanticModel) {
-        ServiceDatabaseManager databaseManager = ServiceDatabaseManager.getInstance();
-        List<Annotation> annotationAttachments = databaseManager.
-                getAnnotationAttachments("ballerina", "http", "OBJECT_METHOD");
-        Map<String, Value> annotations = Function.createAnnotationsMap(annotationAttachments);
-        Service serviceModel = Service.getNewService();
-        List<Function> functionModels = new ArrayList<>();
-        serviceDeclarationNode.members().forEach(member -> {
-            if (member instanceof FunctionDefinitionNode functionDefinitionNode) {
-                Function functionModel = getFunctionModel(functionDefinitionNode, semanticModel, true, false,
-                        annotations);
-                functionModel.setEditable(true);
-                functionModels.add(functionModel);
-            }
-        });
-        serviceModel.setFunctions(functionModels);
-        return serviceModel;
     }
 
     public static Optional<Symbol> getHttpServiceContractSym(SemanticModel semanticModel,
@@ -523,74 +465,6 @@ public final class Utils {
         return Optional.empty();
     }
 
-    public static void updateHttpServiceContractModel(Service serviceModel, TypeDefinitionNode serviceTypeNode,
-                                                      ServiceDeclarationNode serviceDeclaration,
-                                                      SemanticModel semanticModel) {
-        Service commonSvcModel = fromHttpServiceWithContract(serviceTypeNode, serviceDeclaration, semanticModel);
-        enableContractFirstApproach(serviceModel);
-        updateServiceInfo(serviceModel, commonSvcModel);
-        serviceModel.setCodedata(new Codedata(serviceDeclaration.lineRange()));
-        populateListenerInfo(serviceModel, serviceDeclaration);
-    }
-
-    public static void updateHttpServiceModel(Service serviceModel, ServiceDeclarationNode serviceNode,
-                                              SemanticModel semanticModel) {
-        Service commonSvcModel = getServiceModel(serviceNode, semanticModel);
-        updateServiceInfo(serviceModel, commonSvcModel);
-        serviceModel.setCodedata(new Codedata(serviceNode.lineRange()));
-        populateListenerInfo(serviceModel, serviceNode);
-        updateAnnotationAttachmentProperty(serviceNode, serviceModel);
-
-        // handle base path and string literal
-        String attachPoint = getPath(serviceNode.absoluteResourcePath());
-        if (!attachPoint.isEmpty()) {
-            Value basePathProperty = serviceModel.getBasePath();
-            if (Objects.nonNull(basePathProperty)) {
-                basePathProperty.setValue(attachPoint);
-            } else {
-                serviceModel.setBasePath(ServiceModelUtils.getBasePathProperty(attachPoint));
-            }
-        }
-    }
-
-    private static void updateServiceInfo(Service serviceModel, Service commonSvcModel) {
-        populateRequiredFuncsDesignApproachAndServiceType(serviceModel);
-        updateValue(serviceModel.getServiceContractTypeNameValue(), commonSvcModel.getServiceContractTypeNameValue());
-
-        // functions contains in source but not enforced using the service contract type
-        commonSvcModel.getFunctions().forEach(functionModel -> {
-                if (functionModel.getKind().equals(ServiceModelGeneratorConstants.KIND_RESOURCE)) {
-                    getResourceFunctionModel().ifPresentOrElse(
-                            resourceFunction -> {
-                                // remove the default json response from the resource function
-                                if (resourceFunction.getReturnType().getResponses().size() > 1) {
-                                    resourceFunction.getReturnType().getResponses().remove(1);
-                                }
-                                updateFunctionInfo(resourceFunction, functionModel);
-                                serviceModel.addFunction(resourceFunction);
-                            },
-                            () -> serviceModel.addFunction(functionModel)
-                    );
-                } else {
-                    serviceModel.addFunction(functionModel);
-                }
-        });
-    }
-
-    public static Optional<Function> getResourceFunctionModel() {
-        InputStream resourceStream = Utils.class.getClassLoader()
-                .getResourceAsStream("functions/http_resource.json");
-        if (resourceStream == null) {
-            return Optional.empty();
-        }
-
-        try (JsonReader reader = new JsonReader(new InputStreamReader(resourceStream, StandardCharsets.UTF_8))) {
-            return Optional.of(new Gson().fromJson(reader, Function.class));
-        } catch (IOException e) {
-            return Optional.empty();
-        }
-    }
-
     public static Optional<Function> getFunctionModel(String serviceType, String functionNameOrType) {
         String resourcePath =  String.format("functions/%s_%s.json", serviceType.toLowerCase(Locale.US),
                 functionNameOrType.toLowerCase(Locale.US));
@@ -605,21 +479,6 @@ public final class Utils {
         } catch (IOException e) {
             return Optional.empty();
         }
-    }
-
-    public static void updateFunctionInfo(Function functionModel, Function commonFunction) {
-        functionModel.setEditable(commonFunction.isEditable());
-        functionModel.setEnabled(true);
-        functionModel.setKind(commonFunction.getKind());
-        functionModel.setCodedata(commonFunction.getCodedata());
-        updateValue(functionModel.getAccessor(), commonFunction.getAccessor());
-        updateValue(functionModel.getName(), commonFunction.getName());
-        updateValue(functionModel.getReturnType(), commonFunction.getReturnType());
-        List<Parameter> parameters = functionModel.getParameters();
-        parameters.removeIf(parameter -> commonFunction.getParameters().stream()
-                .anyMatch(newParameter -> newParameter.getType().getValue()
-                        .equals(parameter.getType().getValue())));
-        commonFunction.getParameters().forEach(functionModel::addParameter);
     }
 
     public static void populateListenerInfo(Service serviceModel, ServiceDeclarationNode serviceNode) {
